Nearly complete int/long -> size_t/Py_ssize_t conversion.

+++ MomentumAdvection.pxd
@@ -6,8 +6,8 @@ cimport ReferenceState
 cdef class MomentumAdvection:
     cdef:
         double [:] flux
-        long order
+        Py_ssize_t order
 
     cpdef initialize(self,Grid.Grid Gr, PrognosticVariables.PrognosticVariables PV)
     cpdef update(self, Grid.Grid Gr,ReferenceState.ReferenceState Rs, PrognosticVariables.PrognosticVariables PV, ParallelMPI.ParallelMPI Pa)
-    cpdef double [:,:,:] get_flux(self, int i_advected, int i_advecting, Grid.Grid Gr)
+    cpdef double [:,:,:] get_flux(self, Py_ssize_t i_advected, Py_ssize_t i_advecting, Grid.Grid Gr)
