+++ PressureFFTSerial.pxd
@@ -10,22 +10,11 @@ cdef class PressureFFTSerial:
         double [:] a
         double [:] b
         double [:] c
-
         double [:] kx2
         double [:] ky2
-
-
-
-
-        inline void compute_diagonal(self, Grid.Grid Gr, ReferenceState.ReferenceState RS, long i, long j) nogil
-
+        inline void compute_diagonal(self, Grid.Grid Gr, ReferenceState.ReferenceState RS, Py_ssize_t i, Py_ssize_t j) nogil
     cdef SparseSolvers.TDMA TDMA_Solver
-
-
     cpdef initialize(self, Grid.Grid Gr, ReferenceState.ReferenceState RS, ParallelMPI.ParallelMPI Pa)
-
     cpdef compute_modified_wave_numbers(self,Grid.Grid Gr)
-
     cpdef compute_off_diagonals(self,Grid.Grid Gr, ReferenceState.ReferenceState RS)
-
     cpdef solve(self,Grid.Grid Gr, ReferenceState.ReferenceState RS,DiagnosticVariables.DiagnosticVariables DV, ParallelMPI.ParallelMPI PM)
