+++ PrognosticVariables.pxd
@@ -4,23 +4,23 @@ cimport ParallelMPI
 
 cdef extern from "prognostic_variables.h":
     struct VelocityDofs:
-        int u
-        int v
-        int w
+        Py_ssize_t u
+        Py_ssize_t v
+        Py_ssize_t w
 
 cdef extern from "prognostic_variables.h":
-        void build_buffer(int nv, int dim, int s ,Grid.DimStruct *dims, double* values, double* buffer)
-        void buffer_to_values(int dim, int s, Grid.DimStruct *dims, double* values, double* buffer)
-        void set_bcs(int dim, int s, double bc_factor,  Grid.DimStruct *dims, double* values)
-        void set_to_zero(int nv, Grid.DimStruct *dims, double* values )
+        void build_buffer(Py_ssize_t nv, Py_ssize_t dim, Py_ssize_t s ,Grid.DimStruct *dims, double* values, double* buffer)
+        void buffer_to_values(Py_ssize_t dim, Py_ssize_t s, Grid.DimStruct *dims, double* values, double* buffer)
+        void set_bcs(Py_ssize_t dim, Py_ssize_t s, double bc_factor,  Grid.DimStruct *dims, double* values)
+        void set_to_zero(Py_ssize_t nv, Grid.DimStruct *dims, double* values )
 
 cdef class PrognosticVariables:
     cdef:
         dict name_index
         dict units
-        int nv
-        int nv_scalars
-        int nv_velocities
+        Py_ssize_t nv
+        Py_ssize_t nv_scalars
+        Py_ssize_t nv_velocities
         cdef double [:] bc_type
         cdef long [:] var_type
         cdef double [:] values
@@ -32,10 +32,10 @@ cdef class PrognosticVariables:
     cdef:
         void update_all_bcs(self, Grid.Grid Gr, ParallelMPI.ParallelMPI Pa)
     cpdef Update_all_bcs(self,Grid.Grid Gr, ParallelMPI.ParallelMPI Pa)
-    cpdef set_velocity_direction(self,name,int direction,ParallelMPI.ParallelMPI Pa)
-    cdef inline int get_nv(self, str variable_name):
+    cpdef set_velocity_direction(self,name,Py_ssize_t direction,ParallelMPI.ParallelMPI Pa)
+    cdef inline Py_ssize_t get_nv(self, str variable_name):
         return self.name_index[variable_name]
-    cdef inline int get_varshift(self, Grid.Grid Gr, str variable_name):
+    cdef inline Py_ssize_t get_varshift(self, Grid.Grid Gr, str variable_name):
         return self.name_index[variable_name] * Gr.dims.npg
     cpdef get_variable_array(self,name,Grid.Grid Gr)
     cpdef get_tendency_array(self,name,Grid.Grid Gr)
